+++ server/lib/trpc.ts
@@ -1,33 +1,9 @@
 import { initTRPC, TRPCError, type inferProcedureBuilderResolverOptions } from '@trpc/server';
-import type { FetchCreateContextFnOptions } from '@trpc/server/adapters/fetch';
 import { DrizzleQueryError } from 'drizzle-orm/errors';
-import { parseCookie, type CookieHeaders } from './CookieHeaders';
-import sessions from './sessions';
+import type { Context } from './itty';
 import { $ZodError } from 'zod/v4/core';
 import type { DefaultErrorShape } from '@trpc/server/unstable-core-do-not-import';
-import { createDatabase } from './db';
 import ErrorCodes from './ErrorCodes';
-
-export function createTrpcContextFactory(env: Env, ctx: ExecutionContext, resHeaders: CookieHeaders) {
-  const db = createDatabase(env);
-  return async function ({ req }: FetchCreateContextFnOptions) {
-    const reqCookie = parseCookie(req);
-    const checkResult = await sessions.check(db, req, env, resHeaders, reqCookie);
-
-    return {
-      db,
-      req,
-      env, // Cloudflare workers enviroment
-      wctx: ctx, // Cloudflare workers context
-      url: new URL(req.url),
-      resHeaders,
-      ...checkResult,
-      reqCookie,
-    };
-  };
-}
-
-export type Context = Awaited<ReturnType<ReturnType<typeof createTrpcContextFactory>>>;
 
 export type AppErrorShapeData = DefaultErrorShape['data'] & {
   fieldErrors?: Record<string, string[] | undefined>;
